Updated the homepage 1st round
Got the GIF image as the jumbotron background, had to mess with it to
get it positioned and sized correctly and tried to bring in a google
font for the GIF Canyon title. Got rid of the mean.js text and replaced
it with text about our site. Commented out the other unneeded sections
and copied the code from the learn more button and made them create and
browse. I tried to change the nav bar drop downs (which only appear
after you are signed in) to Browse>Categories and Liked GIF's and My
GIFs> View GIFs and Create a GIF. I changed the titles on the page when
you click on the drop downs but the nav bar it's self is a little more
complicated to change because it broke when I tried. I can't figure out
how to change the title of the site that appears in the tab with the
favicon. It broke the site when I tried changing it.